refactor: архивирование легаси UserManager и переименование V2 в основную версию
- Архивированы старые версии UserManager и ApprovedUsersStorage в папку archived/
- UserManagerV2 переименован в UserManager
- ApprovedUsersStorageV2 переименован в ApprovedUsersStorage
- Удалена условная логика Config.UseNewApprovalSystem
- Обновлены все тесты и зависимости
- Упрощена логика в Worker и ModerationService
- Все 593 теста проходят успешно
