+++ archived/ApprovedUsersStorageTestFactory_legacy.cs
@@ -7,7 +7,7 @@ using NUnit.Framework;
 namespace ClubDoorman.TestInfrastructure;
 
 /// <summary>
-/// TestFactory для ApprovedUsersStorageV2
+/// TestFactory для ApprovedUsersStorage
 /// Автоматически сгенерировано
 /// </summary>
 [TestFixture]
@@ -21,6 +21,19 @@ public class ApprovedUsersStorageTestFactory
         return new ApprovedUsersStorage(
             LoggerMock.Object
         );
+    }
+
+    /// <summary>
+    /// Создает экземпляр с кастомным путем к файлу для тестирования
+    /// </summary>
+    public ApprovedUsersStorage CreateApprovedUsersStorageWithCustomPath(string filePath)
+    {
+        // Используем рефлексию для установки приватного поля _filePath
+        var service = new ApprovedUsersStorage(LoggerMock.Object);
+        var field = typeof(ApprovedUsersStorage).GetField("_filePath", 
+            System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+        field?.SetValue(service, filePath);
+        return service;
     }
 
     #region Configuration Methods
